Update viewer to support both normal methods for point cloud and poisson export
