+++ nerfstudio/viewer/app/src/modules/SidePanel/ExportPanel/MeshSubPanel.jsx
@@ -6,23 +6,7 @@ import ContentCopyRoundedIcon from '@mui/icons-material/ContentCopyRounded';
 
 import { Button } from '@mui/material';
 
-function get_normal_outputs(output_options) {
-  // Get a list of normal outputs from the Model
-  let normal_options = [];
-  if (output_options) {
-    // check which outputs have normals
-    for (let i = 0; i < output_options.length; i += 1) {
-      const output_name = output_options[i];
-      if (output_name.includes('normals')) {
-        normal_options.push(output_name);
-      }
-    }
-  }
-  if (normal_options.length === 0) {
-    normal_options = ['none'];
-  }
-  return normal_options;
-}
+import { get_normal_methods } from '../../../utils';
 
 export default function MeshSubPanel(props) {
   const update_box_center = props.update_box_center;
@@ -47,7 +31,7 @@ export default function MeshSubPanel(props) {
   const output_options = useSelector(
     (state) => state.renderingState.output_options,
   );
-  const normal_options = get_normal_outputs(output_options);
+  const normal_options = get_normal_methods(output_options);
   const mesh_method_options = ['poisson', 'tsdf'];
 
   const config_filename = `${config_base_dir}/config.yml`;
@@ -64,8 +48,11 @@ export default function MeshSubPanel(props) {
       },
       normal_options: {
         label: 'Normal Options',
-        options: [...new Set(normal_options)],
-        value: normal_options[0],
+        options: [...normal_options],
+        value:
+          normal_options.values().next().value !== 'none'
+            ? normal_options.values().next().value
+            : 'open3d',
         render: (get) => get('mesh_method_options') === 'poisson',
       },
       numFaces: { label: 'Number of Faces', value: 50000, min: 1 },
@@ -162,7 +149,12 @@ export default function MeshSubPanel(props) {
       ` --output-dir ${controlValues.outputDir}` +
       ` --target-num-faces ${controlValues.numFaces}` +
       ` --num-pixels-per-side ${controlValues.textureResolution}` +
-      ` --normal-output-name ${controlValues.normal_options}` +
+      ` --normal-method ${
+        controlValues.normal_options === 'open3d' ? 'open3d' : 'model_output'
+      }` +
+      (controlValues.normal_options !== 'open3d'
+        ? ` --normal-output-name ${controlValues.normal_options}`
+        : '') +
       ` --num-points ${controlValues.numPoints}` +
       ` --remove-outliers ${getPythonBool(controlValues.removeOutliers)}` +
       ` --use-bounding-box ${getPythonBool(clippingEnabled)}` +
